Fix using double batch size when testing
Since half the batch size goes to labeled and half to unlabeled. If minimum
occurrence is set to guarantee there's enough instances for the half batch
size, using the full batch size for both groups can cause some genes to not
have enough samples, leaving 0s at the end of the similarity array.