Collect all annotations instead of only genes
Major rework of dataset creation. Break up reading from the archives with
attaching foreign data.

- Collect labels for all annotations.
- Add titles.
- Remove gene symbols, which weren't handled well before.
- Remove masking abstracts. This can now be done elsewhere so the dataset
produced is generic and can be masked as needed.
- Store a sequence of dictionaries for tracking location in text of each
annotation. Providing the information needed for masking.
- Remove pubmed edges. This should be added elsewhere. Major benefit is
attaching the edges has been flaky and causing errors. Better to return the
initial dataset after the slowest I/O parts then mess with attaching the edges
instead of having to rerun the I/O if it fails at the end of a full run.